% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_functions.R
\docType{package}
\name{cloudmonitoring_googleAuthR}
\alias{cloudmonitoring_googleAuthR}
\alias{cloudmonitoring_googleAuthR-package}
\title{Cloud Monitoring API
Accesses Google Cloud Monitoring data.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2016-09-03 22:59:49
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudmonitoringv2beta2.auto/R/cloudmonitoring_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
}
}

