% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{DeleteMetricDescriptorResponse}
\alias{DeleteMetricDescriptorResponse}
\title{Cloud Monitoring API Objects 
Accesses Google Cloud Monitoring data.}
\usage{
DeleteMetricDescriptorResponse()
}
\value{
DeleteMetricDescriptorResponse object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2016-09-03 22:59:49
filename: /Users/mark/dev/R/autoGoogleAPI/googlecloudmonitoringv2beta2.auto/R/cloudmonitoring_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
DeleteMetricDescriptorResponse Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response of cloudmonitoring.metricDescriptors.delete.
}

