% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{ListTimeseriesDescriptorsResponse}
\alias{ListTimeseriesDescriptorsResponse}
\title{ListTimeseriesDescriptorsResponse Object}
\usage{
ListTimeseriesDescriptorsResponse(nextPageToken = NULL, oldest = NULL,
  timeseries = NULL, youngest = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token}

\item{oldest}{The oldest timestamp of the interval of this query, as an RFC 3339 string}

\item{timeseries}{The returned time series descriptors}

\item{youngest}{The youngest timestamp of the interval of this query, as an RFC 3339 string}
}
\value{
ListTimeseriesDescriptorsResponse object
}
\description{
ListTimeseriesDescriptorsResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response of cloudmonitoring.timeseriesDescriptors.list
}

