% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{ListTimeseriesResponse}
\alias{ListTimeseriesResponse}
\title{ListTimeseriesResponse Object}
\usage{
ListTimeseriesResponse(nextPageToken = NULL, oldest = NULL,
  timeseries = NULL, youngest = NULL)
}
\arguments{
\item{nextPageToken}{Pagination token}

\item{oldest}{The oldest timestamp of the interval of this query as an RFC 3339 string}

\item{timeseries}{The returned time series}

\item{youngest}{The youngest timestamp of the interval of this query as an RFC 3339 string}
}
\value{
ListTimeseriesResponse object
}
\description{
ListTimeseriesResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response of cloudmonitoring.timeseries.list
}
