% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{MetricDescriptorLabelDescriptor}
\alias{MetricDescriptorLabelDescriptor}
\title{MetricDescriptorLabelDescriptor Object}
\usage{
MetricDescriptorLabelDescriptor(description = NULL, key = NULL)
}
\arguments{
\item{description}{Label description}

\item{key}{Label key}
}
\value{
MetricDescriptorLabelDescriptor object
}
\description{
MetricDescriptorLabelDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A label in a metric is a description of this metric, including the key of this description (what the description is), and the value for this description.
}

