% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_functions.R
\name{metricDescriptors.list}
\alias{metricDescriptors.list}
\title{List metric descriptors that match the query. If the query is not set, then all of the metric descriptors will be returned. Large responses will be paginated, use the nextPageToken returned in the response to request subsequent pages of results by setting the pageToken query parameter to the value of the nextPageToken.}
\usage{
metricDescriptors.list(ListMetricDescriptorsRequest, project, count = NULL,
  pageToken = NULL, query = NULL)
}
\arguments{
\item{ListMetricDescriptorsRequest}{The \link{ListMetricDescriptorsRequest} object to pass to this method}

\item{project}{The project id}

\item{count}{Maximum number of metric descriptors per page}

\item{pageToken}{The pagination token, which is used to page through large result sets}

\item{query}{The query used to search against existing metrics}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/monitoring)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/monitoring/v2beta2/}{Google Documentation}

Other ListMetricDescriptorsRequest functions: \code{\link{ListMetricDescriptorsRequest}}
}
