% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{MetricDescriptorTypeDescriptor}
\alias{MetricDescriptorTypeDescriptor}
\title{MetricDescriptorTypeDescriptor Object}
\usage{
MetricDescriptorTypeDescriptor(metricType = NULL, valueType = NULL)
}
\arguments{
\item{metricType}{The method of collecting data for the metric}

\item{valueType}{The data type of of individual points in the metric's time series}
}
\value{
MetricDescriptorTypeDescriptor object
}
\description{
MetricDescriptorTypeDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A type in a metric contains information about how the metric is collected and what its data points look like.
}

