% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{Point}
\alias{Point}
\title{Point Object}
\usage{
Point(boolValue = NULL, distributionValue = NULL, doubleValue = NULL,
  end = NULL, int64Value = NULL, start = NULL, stringValue = NULL)
}
\arguments{
\item{boolValue}{The value of this data point}

\item{distributionValue}{The value of this data point as a distribution}

\item{doubleValue}{The value of this data point as a double-precision floating-point number}

\item{end}{The interval [start, end] is the time period to which the point's value applies}

\item{int64Value}{The value of this data point as a 64-bit integer}

\item{start}{The interval [start, end] is the time period to which the point's value applies}

\item{stringValue}{The value of this data point in string format}
}
\value{
Point object
}
\description{
Point Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Point is a single point in a time series. It consists of a start time, an end time, and a value.
}
