% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{PointDistribution}
\alias{PointDistribution}
\title{PointDistribution Object}
\usage{
PointDistribution(buckets = NULL, overflowBucket = NULL,
  underflowBucket = NULL)
}
\arguments{
\item{buckets}{The finite buckets}

\item{overflowBucket}{The overflow bucket}

\item{underflowBucket}{The underflow bucket}
}
\value{
PointDistribution object
}
\description{
PointDistribution Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Distribution data point value type. When writing distribution points, try to be consistent with the boundaries of your buckets. If you must modify the bucket boundaries, then do so by merging, partitioning, or appending rather than skewing them.
}
