% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{PointDistributionBucket}
\alias{PointDistributionBucket}
\title{PointDistributionBucket Object}
\usage{
PointDistributionBucket(count = NULL, lowerBound = NULL,
  upperBound = NULL)
}
\arguments{
\item{count}{The number of events whose values are in the interval defined by this bucket}

\item{lowerBound}{The lower bound of the value interval of this bucket (inclusive)}

\item{upperBound}{The upper bound of the value interval of this bucket (exclusive)}
}
\value{
PointDistributionBucket object
}
\description{
PointDistributionBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The histogram's bucket. Buckets that form the histogram of a distribution value. If the upper bound of a bucket, say U1, does not equal the lower bound of the next bucket, say L2, this means that there is no event in [U1, L2).
}

