% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{PointDistributionOverflowBucket}
\alias{PointDistributionOverflowBucket}
\title{PointDistributionOverflowBucket Object}
\usage{
PointDistributionOverflowBucket(count = NULL, lowerBound = NULL)
}
\arguments{
\item{count}{The number of events whose values are in the interval defined by this bucket}

\item{lowerBound}{The lower bound of the value interval of this bucket (inclusive)}
}
\value{
PointDistributionOverflowBucket object
}
\description{
PointDistributionOverflowBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The overflow bucket is a special bucket that does not have the upperBound field; it includes all of the events that are no less than its lower bound.
}
