% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{PointDistributionUnderflowBucket}
\alias{PointDistributionUnderflowBucket}
\title{PointDistributionUnderflowBucket Object}
\usage{
PointDistributionUnderflowBucket(count = NULL, upperBound = NULL)
}
\arguments{
\item{count}{The number of events whose values are in the interval defined by this bucket}

\item{upperBound}{The upper bound of the value interval of this bucket (exclusive)}
}
\value{
PointDistributionUnderflowBucket object
}
\description{
PointDistributionUnderflowBucket Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The underflow bucket is a special bucket that does not have the lowerBound field; it includes all of the events that are less than its upper bound.
}
