% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{Timeseries}
\alias{Timeseries}
\title{Timeseries Object}
\usage{
Timeseries(points = NULL, timeseriesDesc = NULL)
}
\arguments{
\item{points}{The data points of this time series}

\item{timeseriesDesc}{The descriptor of this time series}
}
\value{
Timeseries object
}
\description{
Timeseries Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The monitoring data is organized as metrics and stored as data points that are recorded over time. Each data point represents information like the CPU utilization of your virtual machine. A historical record of these data points is called a time series.
}

