% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_functions.R
\name{timeseries.write}
\alias{timeseries.write}
\title{Put data points to one or more time series for one or more metrics. If a time series does not exist, a new time series will be created. It is not allowed to write a time series point that is older than the existing youngest point of that time series. Points that are older than the existing youngest point of that time series will be discarded silently. Therefore, users should make sure that points of a time series are written sequentially in the order of their end time.}
\usage{
timeseries.write(WriteTimeseriesRequest, project)
}
\arguments{
\item{WriteTimeseriesRequest}{The \link{WriteTimeseriesRequest} object to pass to this method}

\item{project}{The project ID}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/cloud-platform
\item https://www.googleapis.com/auth/monitoring
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/cloud-platform, https://www.googleapis.com/auth/monitoring)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://cloud.google.com/monitoring/v2beta2/}{Google Documentation}

Other WriteTimeseriesRequest functions: \code{\link{WriteTimeseriesRequest.commonLabels}},
  \code{\link{WriteTimeseriesRequest}}
}
