% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{TimeseriesDescriptor.labels}
\alias{TimeseriesDescriptor.labels}
\title{TimeseriesDescriptor.labels Object}
\usage{
TimeseriesDescriptor.labels()
}
\value{
TimeseriesDescriptor.labels object
}
\description{
TimeseriesDescriptor.labels Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The label's name.
}
\seealso{
Other TimeseriesDescriptor functions: \code{\link{TimeseriesDescriptor}}
}

