% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{TimeseriesDescriptor}
\alias{TimeseriesDescriptor}
\title{TimeseriesDescriptor Object}
\usage{
TimeseriesDescriptor(TimeseriesDescriptor.labels = NULL, labels = NULL,
  metric = NULL, project = NULL)
}
\arguments{
\item{TimeseriesDescriptor.labels}{The \link{TimeseriesDescriptor.labels} object or list of objects}

\item{labels}{The label's name}

\item{metric}{The name of the metric}

\item{project}{The Developers Console project number to which this time series belongs}
}
\value{
TimeseriesDescriptor object
}
\description{
TimeseriesDescriptor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
TimeseriesDescriptor identifies a single time series.
}
\seealso{
Other TimeseriesDescriptor functions: \code{\link{TimeseriesDescriptor.labels}}
}
