% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{TimeseriesDescriptorLabel}
\alias{TimeseriesDescriptorLabel}
\title{TimeseriesDescriptorLabel Object}
\usage{
TimeseriesDescriptorLabel(key = NULL, value = NULL)
}
\arguments{
\item{key}{The label's name}

\item{value}{The label's value}
}
\value{
TimeseriesDescriptorLabel object
}
\description{
TimeseriesDescriptorLabel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
No description
}

