% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{TimeseriesPoint}
\alias{TimeseriesPoint}
\title{TimeseriesPoint Object}
\usage{
TimeseriesPoint(point = NULL, timeseriesDesc = NULL)
}
\arguments{
\item{point}{The data point in this time series snapshot}

\item{timeseriesDesc}{The descriptor of this time series}
}
\value{
TimeseriesPoint object
}
\description{
TimeseriesPoint Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
When writing time series, TimeseriesPoint should be used instead of Timeseries, to enforce single point for each time series in the timeseries.write request.
}
