% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloudmonitoring_objects.R
\name{WriteTimeseriesRequest}
\alias{WriteTimeseriesRequest}
\title{WriteTimeseriesRequest Object}
\usage{
WriteTimeseriesRequest(WriteTimeseriesRequest.commonLabels = NULL,
  commonLabels = NULL, timeseries = NULL)
}
\arguments{
\item{WriteTimeseriesRequest.commonLabels}{The \link{WriteTimeseriesRequest.commonLabels} object or list of objects}

\item{commonLabels}{The label's name}

\item{timeseries}{Provide time series specific labels and the data points for each time series}
}
\value{
WriteTimeseriesRequest object
}
\description{
WriteTimeseriesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The request of cloudmonitoring.timeseries.write
}
\seealso{
Other WriteTimeseriesRequest functions: \code{\link{WriteTimeseriesRequest.commonLabels}},
  \code{\link{timeseries.write}}
}

